fix: move  avatarURL to constant.js & timAgo func to utils.
