+++ src/utils/constants.js
@@ -1,2 +1,4 @@
 export const baseUrl = "https://backend-productivity-graveyard.onrender.com";
+export const avatarApiUrl = "https://ui-avatars.com/api";
+
 // export const baseUrl = "http://localhost:3000";
